+++ src/components/nativevideo/NativeVideoConfig.ts
@@ -1,0 +1,13 @@
+/******************************************************************
+ * NativeVideoConfig
+ *
+ * @author matthias.schulz@jash.de
+ *****************************************************************/
+
+export interface NativeVideoConfig {
+    muted?: boolean;
+    preload?: boolean;
+    loop?: boolean;
+    controls?: boolean;
+    autoplay?: boolean;
+}