+++ src/components/nativevideo/NativeVideoSource.ts
@@ -1,0 +1,10 @@
+/******************************************************************
+ * NativeVideoSource
+ *
+ * @author matthias.schulz@jash.de
+ *****************************************************************/
+
+export interface NativeVideoSource {
+    videoURL: string;
+    videoType?: string;
+}